More work on multipart
Guts done. Still todo:

- clean up temp files that are created, possibly using CPS (want to
  avoid the way Ring does it with TTL and reaper thread)
- clean up parseFileItem and friends -- get rid of `.get` calls

+++ src/main/scala/com/markfeeney/circlet/middleware/Param.scala
@@ -23,7 +23,7 @@ final case class FileParam(
   filename: String,
   contentType: String,
   tempFile: File,
-  size: Int) extends Param
+  size: Long) extends Param
 
 object Param {
   implicit def str2StrParam(s: String): StrParam = StrParam(Vector(s))
